fix: sort stable version
