+++ crates/pop-parachains/src/up/relay.rs
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0
 
 use super::chain_specs::chain_spec_generator;
+use pop_common::helpers::parse_latest_tag;
 pub use pop_common::{
 	git::GitHub,
 	sourcing::{
@@ -104,7 +105,10 @@ pub(super) async fn from(
 		let releases = relay.releases().await?;
 		let tag = Binary::resolve_version(name, version, &releases, cache);
 		// Only set latest when caller has not explicitly specified a version to use
-		let latest = version.is_none().then(|| releases.first().map(|v| v.to_string())).flatten();
+		let latest = version
+			.is_none()
+			.then(|| parse_latest_tag(releases.iter().map(|s| s.as_str()).collect()))
+			.flatten();
 		let binary = Binary::Source {
 			name: name.to_string(),
 			source: TryInto::try_into(&relay, tag, latest)?,
